Moves test folder outside of src fso coverage does not include testing for testing
