+++ Pipfile
@@ -21,6 +21,7 @@ pytest-cov = "*"
 fastapi = "*"
 uvicorn = "*"
 gunicorn = "*"
+psycopg2 = "*"
 
 [requires]
 python_version = "3.7"
